+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/window/DesktopPopup.desktop.kt
@@ -135,88 +135,6 @@ fun Popup(
     )
 }
 
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-private fun PopupLayout(
-    popupPositionProvider: PopupPositionProvider,
-    focusable: Boolean,
-    onDismissRequest: (() -> Unit)?,
-    onPreviewKeyEvent: ((KeyEvent) -> Boolean) = { false },
-    onKeyEvent: ((KeyEvent) -> Boolean) = { false },
-    content: @Composable () -> Unit
-) {
-    val scene = LocalComposeScene.current
-    val density = LocalDensity.current
-
-    var parentBounds by remember { mutableStateOf(IntRect.Zero) }
-
-    // getting parent bounds
-    Layout(
-        content = {},
-        modifier = Modifier.onGloballyPositioned { childCoordinates ->
-            val coordinates = childCoordinates.parentCoordinates!!
-            parentBounds = IntRect(
-                coordinates.localToWindow(Offset.Zero).round(),
-                coordinates.size
-            )
-        },
-        measurePolicy = { _, _ ->
-            layout(0, 0) {}
-        }
-    )
-
-    val parentComposition = rememberCompositionContext()
-    val (owner, composition) = remember {
-        val owner = SkiaBasedOwner(
-            scene = scene,
-            platform = scene.platform,
-            pointerPositionUpdater = scene.pointerPositionUpdater,
-            density = density,
-            isFocusable = focusable,
-            onDismissRequest = onDismissRequest,
-            onPreviewKeyEvent = onPreviewKeyEvent,
-            onKeyEvent = onKeyEvent
-        )
-        scene.attach(owner)
-
-        val composition = owner.setContent(parent = parentComposition) {
-            Layout(
-                content = content,
-                measurePolicy = { measurables, constraints ->
-                    val width = constraints.maxWidth
-                    val height = constraints.maxHeight
-
-                    layout(constraints.maxWidth, constraints.maxHeight) {
-                        measurables.forEach {
-                            val placeable = it.measure(constraints)
-                            val position = popupPositionProvider.calculatePosition(
-                                anchorBounds = parentBounds,
-                                windowSize = IntSize(width, height),
-                                layoutDirection = layoutDirection,
-                                popupContentSize = IntSize(placeable.width, placeable.height)
-                            )
-                            owner.bounds = IntRect(
-                                position,
-                                IntSize(placeable.width, placeable.height)
-                            )
-                            placeable.place(position.x, position.y)
-                        }
-                    }
-                }
-            )
-        }
-        owner to composition
-    }
-    owner.density = density
-    DisposableEffect(Unit) {
-        onDispose {
-            scene.detach(owner)
-            composition.dispose()
-            owner.dispose()
-        }
-    }
-}
-
 /**
  * Provides [PopupPositionProvider] relative to the current mouse cursor position.
  *
